community-search-rag/pyproject.toml
[project]
name = "flare-ai-rag"
version = "0.1.0"
description = "Flare AI Kit template for Retrieval-Augmented Generation (RAG)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.1",
    "fastapi>=0.115.8",
    "google-generativeai>=0.8.4",
    "httpx>=0.28.1",
    "openrouter>=1.0",
    "pandas>=2.2.3",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "pyopenssl>=25.0.0",
    "pinecone>=3.0.0",
    "structlog>=25.1.0",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.393",
    "ruff>=0.9.4",
]

[project.scripts]
start-backend = "flare_ai_rag.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D212", "COM812", "D", "S105", "ANN401", "ISC003"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = ["S101", "ARG"]
"src/flare_ai_rag/router/prompts.py" = ["E501"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
pythonVersion = "3.12"
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
deprecateTypingAliases = true
disableBytesTypePromotions = true
reportUnusedClass = true
reportUnusedFunction = true
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportPrivateUsage = true
reportTypeCommentUsage = true
reportConstantRedefinition = true
reportDeprecated = true
reportInconsistentConstructor = true
reportUninitializedInstanceVariable = true
reportMissingParameterType = true
reportCallInDefaultInitializer = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportInvalidStubStatement = true
reportUnusedExpression = true
reportUnnecessaryTypeIgnoreComment = true
reportMatchNotExhaustive = true
reportImplicitOverride = true
reportShadowedImports = true
